fix: HardwareSession sendCmd, recvResponse, setPC , memoryCorruptionTest and Session create
- sendCmd: handle partial writes, guard sockfd < 0, use MSG_NOSIGNAL

- recvResponse: reset timeval inside loop, distinguish r==0 from r<0

- setPC: add recvResponse after resume, error propagation, uint16_t -> uint32_t

- memoryCorruptionTest: check all sendCmd return values, always resume on

  failure, fix OpenOCD regex, improve stderr logging with hex formatting

+++ QemuSession.cpp
@@ -134,7 +134,7 @@ bool QEMUSession::memoryCorruptionTest(uint32_t addr,
 // To mirror HardwareSession::setPC we accept only a uint16_t new PC value.
 // The plugin will redirect execution when the instruction counter fires.
 // Because setPC has no "success range" concept we always return 0.
-int QEMUSession::setPC(uint16_t newPC) {
+int QEMUSession::setPC(uint32_t newPC) {
     // Default: inject at instruction 1 (immediately)
     // Caller can use the extended runCampaign path if a specific count matters
     std::ostringstream js;
